// src/interfaces/planets_type.ts
export default interface Planets {
    planet_data : {
        "name": string
          "diameter": string
          "rotation_period": string
          "orbital_period": string
          "gravity": string
          "population": string
          "climate": string
          "terrain": string
          "surface_water": string
          "films": string[]
          "residents": string[]
          "url": string
          "created": string
          "edited": string
        }
    }